combine options
